decoding for decoder attn mask|debug 3

+++ transformers/modeling_encoder_decoder.py
@@ -305,8 +305,8 @@ class PreTrainedEncoderDecoder(nn.Module):
 
 
         for step in range(10):
-            produced_decoder_attn_mask=torch.cat([torch.ones([decoder_input_shape[0],step+1], device=decoder_input_ids.device)
-                                                     ,torch.zeros([decoder_input_shape[0],decoder_input_shape[1]-(step+1)], device=decoder_input_ids.device)],dim=1)
+            produced_decoder_attn_mask=torch.cat([torch.ones([decoder_input_shape[0],step+1],dtype=torch.int32, device=decoder_input_ids.device)
+                                                     ,torch.zeros([decoder_input_shape[0],decoder_input_shape[1]-(step+1)], dtype=torch.int32, device=decoder_input_ids.device)],dim=1)
             print('produced_decoder_attn_mask = {}'.format(produced_decoder_attn_mask))
 
             kwargs_decoder["attention_mask"]=produced_decoder_attn_mask
